Split config into separate file
